# OpenC2_Orchestrator/codegen_server/openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  description: API for creating an OpenC2 command as a Producer to deliver it on an
    OpenC2 Consumer
  title: OpenC2 Producer Command Creation App
  version: 0.0.1
servers:
- description: OpenC2 API
  url: http://localhost:8888/
paths:
  /close_connection:
    post:
      description: Send an OpenC2 command to consumer to close a TCP connection
      operationId: close_connection_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseConnInfo'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenC2Response'
          description: Command successfully executed. Connection closed.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
          description: Unexpected error
      summary: Close a TCP connection on OpenC2 consumer/actuator
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /delete_file:
    post:
      description: Send an OpenC2 command to consumer to delete a file
      operationId: delete_file_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteFileInfo'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenC2Response'
          description: Command successfully executed. File deleted.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
          description: Unexpected error
      summary: Delete a file on OpenC2 consumer/actuator
      x-openapi-router-controller: openapi_server.controllers.default_controller
components:
  schemas:
    DeleteFileInfo:
      example:
        file_path: file_path
        actuator_ip: actuator_ip
      properties:
        actuator_ip:
          description: IP of OpenC2 Actuator/Consumer of command
          title: actuator_ip
          type: string
        file_path:
          description: File path of file to be deleted
          title: file_path
          type: string
      required:
      - actuator_ip
      - file_path
      title: DeleteFileInfo
      type: object
    CloseConnInfo:
      example:
        src_port: 6
        actuator_ip: actuator_ip
        dst_port: 1
        pid: 0
        dst_ip: dst_ip
      properties:
        actuator_ip:
          description: IP of OpenC2 Actuator/Consumer of command
          title: actuator_ip
          type: string
        pid:
          description: PID of established connection process
          title: pid
          type: integer
        dst_ip:
          description: Dest IP of established connection
          title: dst_ip
          type: string
        src_port:
          description: Source port of established connection
          title: src_port
          type: integer
        dst_port:
          description: Dest port of established connection
          title: dst_port
          type: integer
      required:
      - actuator_ip
      - dst_ip
      - dst_port
      - pid
      - src_port
      title: CloseConnInfo
      type: object
    OpenC2CommandHeaders:
      example:
        created: 0
        from: from
        to: to
        request_id: request_id
      properties:
        request_id:
          description: ID of a OpenC2 Command Request
          title: request_id
          type: string
        created:
          description: OpenC2 Command creation Timestamp
          title: created
          type: integer
        from:
          description: Hostname or IP of OpenC2 sender
          title: from
          type: string
        to:
          description: Hostname or IP of OpenC2 receiver
          title: to
          type: string
      required:
      - created
      - from
      - request_id
      - to
      title: OpenC2CommandHeaders
      type: object
    OpenC2CommandResponse:
      example:
        status_text: status_text
        results: results
        status: 6
      properties:
        status:
          description: Status code of OpenC2 command execution
          title: status
          type: integer
        status_text:
          description: Status text of OpenC2 command execution
          title: status_text
          type: string
        results:
          description: Results of command execution
          title: results
          type: string
      required:
      - results
      - status
      - status_text
      title: OpenC2CommandResponse
      type: object
    OpenC2Response:
      example:
        command_response:
          status_text: status_text
          results: results
          status: 6
        headers:
          created: 0
          from: from
          to: to
          request_id: request_id
      properties:
        headers:
          $ref: '#/components/schemas/OpenC2CommandHeaders'
        command_response:
          $ref: '#/components/schemas/OpenC2CommandResponse'
      required:
      - command_response
      - headers
      title: OpenC2Response
      type: object
    ErrorMessage:
      example:
        error_message: error_message
        error_code: error_code
      properties:
        error_code:
          title: error_code
          type: string
        error_message:
          title: error_message
          type: string
      required:
      - error_code
      - error_message
      title: ErrorMessage
      type: object